refactor: remove enum.hpp and integrate enums into logging, pipeline, and status headers
